--- components/topic-mgt/org.wso2.carbon.identity.topic.management/src/main/java/org/wso2/carbon/identity/topic/management/api/exception/TopicManagementClientException.java
@@ -1,37 +1,0 @@
-/*
- * Copyright (c) 2025, WSO2 LLC. (http://www.wso2.com).
- *
- * WSO2 LLC. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.carbon.identity.topic.management.api.exception;
-
-/**
- * Exception class for client-side errors in topic management.
- */
-public class TopicManagementClientException extends TopicManagementException {
-
-    /**
-     * Constructs a new exception with the specified detail message, description, and error code.
-     *
-     * @param message     The detail message
-     * @param description The description
-     * @param errorCode   The error code
-     */
-    public TopicManagementClientException(String message, String description, String errorCode) {
-
-        super(message, description, errorCode);
-    }
-}
