--- components/topic-mgt/org.wso2.carbon.identity.topic.management/src/test/java/org/wso2/carbon/identity/topic/management/dao/TopicManagementDAOImplTest.java
@@ -1,88 +1,0 @@
-/*
- * Copyright (c) 2025, WSO2 LLC. (http://www.wso2.com).
- *
- * WSO2 LLC. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.carbon.identity.topic.management.dao;
-
-import org.testng.annotations.Test;
-import org.wso2.carbon.identity.common.testng.WithCarbonHome;
-import org.wso2.carbon.identity.common.testng.WithH2Database;
-import org.wso2.carbon.identity.topic.management.api.exception.TopicManagementException;
-import org.wso2.carbon.identity.topic.management.internal.dao.impl.TopicManagementDAOImpl;
-
-import static org.testng.Assert.assertFalse;
-import static org.testng.Assert.assertTrue;
-
-@WithH2Database(files = {"dbscripts/h2.sql"})
-@WithCarbonHome
-public class TopicManagementDAOImplTest {
-
-    private static final String TOPIC = "https://example.com/events/carbon.super";
-    private static final String TOPIC2 = "https://example.com/events/carbon.super2";
-    private static final int TENANT_ID = 1;
-
-    private final TopicManagementDAOImpl topicManagementDAOImpl = new TopicManagementDAOImpl();
-
-    @Test
-    public void testAddTopic() throws TopicManagementException {
-
-        topicManagementDAOImpl.addTopic(TOPIC, TENANT_ID);
-        assertTrue(topicManagementDAOImpl.isTopicExists(TOPIC, TENANT_ID));
-    }
-
-    @Test(dependsOnMethods = {"testAddTopic"})
-    public void testAddTopicDifferentTenants() throws TopicManagementException {
-
-        topicManagementDAOImpl.addTopic(TOPIC, TENANT_ID + 1);
-        assertTrue(topicManagementDAOImpl.isTopicExists(TOPIC, TENANT_ID));
-        assertTrue(topicManagementDAOImpl.isTopicExists(TOPIC, TENANT_ID + 1));
-    }
-
-    @Test(dependsOnMethods = {"testAddTopic"})
-    public void testIsTopicExistsWhenItExists() throws TopicManagementException {
-
-        assertTrue(topicManagementDAOImpl.isTopicExists(TOPIC, TENANT_ID));
-    }
-
-    @Test
-    public void testIsTopicExistsWhenItDoesNotExist() throws TopicManagementException {
-
-        assertFalse(topicManagementDAOImpl.isTopicExists("nonexistent-topic", TENANT_ID));
-    }
-
-    @Test(dependsOnMethods = {"testAddTopic"})
-    public void testDeleteTopic() throws TopicManagementException {
-
-        topicManagementDAOImpl.addTopic(TOPIC2, TENANT_ID);
-        assertTrue(topicManagementDAOImpl.isTopicExists(TOPIC2, TENANT_ID));
-        topicManagementDAOImpl.deleteTopic(TOPIC2, TENANT_ID);
-        assertFalse(topicManagementDAOImpl.isTopicExists(TOPIC2, TENANT_ID));
-    }
-
-    @Test(dependsOnMethods = {"testAddTopic"})
-    public void testDeleteNonExistentTopic() throws TopicManagementException {
-
-        topicManagementDAOImpl.deleteTopic("nonexistent-topic", TENANT_ID);
-        assertFalse(topicManagementDAOImpl.isTopicExists("nonexistent-topic", TENANT_ID));
-    }
-
-    @Test(dependsOnMethods = {"testAddTopic"}, expectedExceptions = TopicManagementException.class)
-    public void testAddDuplicateTopic() throws TopicManagementException {
-
-        topicManagementDAOImpl.addTopic(TOPIC, TENANT_ID);
-    }
-}
