--- components/topic-mgt/org.wso2.carbon.identity.topic.management/src/test/java/org/wso2/carbon/identity/topic/management/service/TopicManagementServiceImplTest.java
@@ -1,176 +1,0 @@
-/*
- * Copyright (c) 2025, WSO2 LLC. (http://www.wso2.com).
- *
- * WSO2 LLC. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.carbon.identity.topic.management.service;
-
-import org.mockito.MockedStatic;
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.BeforeClass;
-import org.testng.annotations.BeforeMethod;
-import org.testng.annotations.Test;
-import org.wso2.carbon.identity.common.testng.WithCarbonHome;
-import org.wso2.carbon.identity.core.util.IdentityTenantUtil;
-import org.wso2.carbon.identity.topic.management.api.exception.TopicManagementException;
-import org.wso2.carbon.identity.topic.management.api.service.TopicManager;
-import org.wso2.carbon.identity.topic.management.internal.component.TopicManagementComponentServiceHolder;
-import org.wso2.carbon.identity.topic.management.internal.dao.TopicManagementDAO;
-import org.wso2.carbon.identity.topic.management.internal.service.impl.TopicManagementServiceImpl;
-
-import java.lang.reflect.Field;
-import java.util.Collections;
-
-import static org.mockito.Mockito.doNothing;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.mockStatic;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-import static org.testng.Assert.assertEquals;
-import static org.testng.Assert.assertFalse;
-import static org.testng.Assert.assertTrue;
-
-@WithCarbonHome
-public class TopicManagementServiceImplTest {
-
-    private TopicManagementServiceImpl topicManagementService;
-    private TopicManagementDAO topicManagementDAO;
-    private TopicManager topicManager;
-    private TopicManagementComponentServiceHolder componentServiceHolderMock;
-    private MockedStatic<IdentityTenantUtil> identityTenantUtilMockedStatic;
-
-    private static final String TENANT_DOMAIN = "carbon.super";
-    private static final int TENANT_ID = 1;
-    private static final String CHANNEL_URI = "example.com/events";
-    private static final String TOPIC = "https://example.com/events/carbon.super";
-    private static final String WEBSUBHUBTOPICMANAGER = "WEBSUBHUBTOPICMANAGER";
-
-    @BeforeClass
-    public void setUpClass() throws Exception {
-
-        topicManagementService = TopicManagementServiceImpl.getInstance();
-
-        identityTenantUtilMockedStatic = mockStatic(IdentityTenantUtil.class);
-        identityTenantUtilMockedStatic.when(() -> IdentityTenantUtil.getTenantId(TENANT_DOMAIN)).thenReturn(TENANT_ID);
-
-        // Mock and inject the singleton holder
-        componentServiceHolderMock = mock(TopicManagementComponentServiceHolder.class);
-        Field instanceField = TopicManagementComponentServiceHolder.class.getDeclaredField("INSTANCE");
-        instanceField.setAccessible(true);
-        instanceField.set(null, componentServiceHolderMock);
-    }
-
-    @AfterClass
-    public void tearDownClass() {
-
-        identityTenantUtilMockedStatic.close();
-    }
-
-    @BeforeMethod
-    public void setUp() throws Exception {
-
-        topicManagementDAO = mock(TopicManagementDAO.class);
-        topicManager = mock(TopicManager.class);
-
-        when(componentServiceHolderMock.getTopicManagers())
-                .thenReturn(Collections.singletonList(topicManager));
-        when(topicManager.getName()).thenReturn(WEBSUBHUBTOPICMANAGER);
-
-        Field daoField = TopicManagementServiceImpl.class.getDeclaredField("topicManagementDAO");
-        daoField.setAccessible(true);
-        daoField.set(topicManagementService, topicManagementDAO);
-    }
-
-    @Test
-    public void testConstructTopic() throws TopicManagementException {
-
-        when(topicManager.constructTopic(CHANNEL_URI, TENANT_DOMAIN)).thenReturn(TOPIC);
-
-        String result = topicManagementService.constructTopic(CHANNEL_URI, TENANT_DOMAIN);
-        assertEquals(result, TOPIC);
-    }
-
-    @Test(expectedExceptions = TopicManagementException.class)
-    public void testConstructTopicWithNullUri() throws TopicManagementException {
-
-        topicManagementService.constructTopic(null, TENANT_DOMAIN);
-    }
-
-    @Test(expectedExceptions = TopicManagementException.class)
-    public void testConstructTopicWithNoManagers() throws TopicManagementException {
-
-        when(componentServiceHolderMock.getTopicManagers()).thenReturn(Collections.emptyList());
-        topicManagementService.constructTopic(CHANNEL_URI, TENANT_DOMAIN);
-    }
-
-    @Test
-    public void testRegisterTopic() throws TopicManagementException {
-
-        doNothing().when(topicManager).registerTopic(TOPIC, TENANT_DOMAIN);
-        doNothing().when(topicManagementDAO).addTopic(TOPIC, TENANT_ID);
-
-        topicManagementService.registerTopic(TOPIC, TENANT_DOMAIN);
-
-        verify(topicManagementDAO).addTopic(TOPIC, TENANT_ID);
-    }
-
-    @Test(expectedExceptions = TopicManagementException.class)
-    public void testRegisterTopicWithNullTopic() throws TopicManagementException {
-
-        topicManagementService.registerTopic(null, TENANT_DOMAIN);
-    }
-
-    @Test
-    public void testDeregisterTopic() throws TopicManagementException {
-
-        doNothing().when(topicManager).deregisterTopic(TOPIC, TENANT_DOMAIN);
-        doNothing().when(topicManagementDAO).deleteTopic(TOPIC, TENANT_ID);
-
-        topicManagementService.deregisterTopic(TOPIC, TENANT_DOMAIN);
-
-        verify(topicManagementDAO).deleteTopic(TOPIC, TENANT_ID);
-    }
-
-    @Test(expectedExceptions = TopicManagementException.class)
-    public void testDeregisterTopicWithNullTopic() throws TopicManagementException {
-
-        topicManagementService.deregisterTopic(null, TENANT_DOMAIN);
-    }
-
-    @Test
-    public void testIsTopicExists() throws TopicManagementException {
-
-        when(topicManagementDAO.isTopicExists(TOPIC, TENANT_ID)).thenReturn(true);
-
-        boolean result = topicManagementService.isTopicExists(TOPIC, TENANT_DOMAIN);
-        assertTrue(result);
-    }
-
-    @Test
-    public void testIsTopicNotExists() throws TopicManagementException {
-
-        when(topicManagementDAO.isTopicExists(TOPIC, TENANT_ID)).thenReturn(false);
-
-        boolean result = topicManagementService.isTopicExists(TOPIC, TENANT_DOMAIN);
-        assertFalse(result);
-    }
-
-    @Test(expectedExceptions = TopicManagementException.class)
-    public void testIsTopicExistsWithNullTopic() throws TopicManagementException {
-
-        topicManagementService.isTopicExists(null, TENANT_DOMAIN);
-    }
-}
