+++ components/webhook-mgt/org.wso2.carbon.identity.topic.management/src/test/resources/dbscripts/h2.sql
@@ -1,0 +1,12 @@
+CREATE TABLE IF NOT EXISTS IDN_WEBHOOK_TOPIC (
+    ID INTEGER NOT NULL AUTO_INCREMENT,
+    TOPIC VARCHAR(255) NOT NULL,
+    CHANNEL_URI VARCHAR(255) NOT NULL,
+    EVENT_PROFILE_VERSION VARCHAR(32) NOT NULL,
+    TENANT_ID INTEGER NOT NULL,
+    PRIMARY KEY (ID),
+    UNIQUE (TOPIC, TENANT_ID)
+    );
+
+-- Indexes
+CREATE INDEX IDX_IDN_WEBHOOK_TOPIC_TOPIC ON IDN_WEBHOOK_TOPIC (TOPIC);
