+++ components/webhook-mgt/org.wso2.carbon.identity.topic.management/src/test/resources/testng.xml
@@ -1,0 +1,38 @@
+<!--
+  ~ Copyright (c) 2025, WSO2 LLC. (http://www.wso2.com).
+  ~
+  ~ WSO2 LLC. licenses this file to you under the Apache License,
+  ~ Version 2.0 (the "License"); you may not use this file except
+  ~ in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~ http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing,
+  ~ software distributed under the License is distributed on an
+  ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+  ~ KIND, either express or implied.  See the License for the
+  ~ specific language governing permissions and limitations
+  ~ under the License.
+  -->
+
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
+
+<suite name="org.wso2.carbon.identity.topic.management">
+    <test name="topic-management-internal-test">
+        <classes>
+            <class name="org.wso2.carbon.identity.topic.management.internal.TopicManagementComponentServiceHolderTest"/>
+        </classes>
+    </test>
+    <test name="topic-management-service-test">
+        <classes>
+            <class name="org.wso2.carbon.identity.topic.management.service.TopicManagementServiceImplTest"/>
+        </classes>
+    </test>
+    <test name="topic-management-dao-test">
+        <classes>
+            <class name="org.wso2.carbon.identity.topic.management.dao.TopicManagementDAOImplTest"/>
+            <class name="org.wso2.carbon.identity.topic.management.dao.CacheBackedTopicManagementDAOTest"/>
+        </classes>
+    </test>
+</suite>
