+++ components/webhook-mgt/pom.xml
@@ -38,5 +38,6 @@
     <modules>
         <module>org.wso2.carbon.identity.webhook.management</module>
         <module>org.wso2.carbon.identity.webhook.metadata</module>
+        <module>org.wso2.carbon.identity.topic.management</module>
     </modules>
 </project>
