+++ features/webhook-mgt/org.wso2.carbon.identity.webhook.management.server.feature/pom.xml
@@ -41,6 +41,10 @@
             <groupId>org.wso2.carbon.identity.framework</groupId>
             <artifactId>org.wso2.carbon.identity.webhook.metadata</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.wso2.carbon.identity.framework</groupId>
+            <artifactId>org.wso2.carbon.identity.topic.management</artifactId>
+        </dependency>
     </dependencies>
 
     <build>
@@ -70,6 +74,9 @@
                                 </bundleDef>
                                 <bundleDef>
                                     org.wso2.carbon.identity.framework:org.wso2.carbon.identity.webhook.metadata
+                                </bundleDef>
+                                <bundleDef>
+                                    org.wso2.carbon.identity.framework:org.wso2.carbon.identity.topic.management
                                 </bundleDef>
                             </bundles>
                         </configuration>
